First replace of files paths ; fix heritage ; fix namespaces ; light refactor of controller and entity
