+++ app/Views/list.twig
@@ -3,11 +3,11 @@
 {% block view %}
 	{% for post in posts %}
 		<section style="width: 850px; margin: auto">
-			<h3>{{ post.titre }}, <em>{{ post.dateAjout }}</em></h3>
+			<h3>{{ post.title }}, <em>{{ post.date }}</em></h3>
 			<img src="{{ post.image }}" style="display:block; margin: auto">
-			<p>{{ post.extrait }}</p>
+			<p>{{ post.sum }}</p>
 			<nav>
-				<a href="{{ post.url }}">Lire {{ post.titre }}</a>	
+				<a href="{{ post.url }}">Lire {{ post.title }}</a>
 			</nav>
 		</section>
 	{% endfor %}
