+++ app/Views/show.twig
@@ -2,9 +2,9 @@
 
 {% block view %}
 
-	<h1>{{ post._titre }}</h1>
-	<span>Posté le {{ post.dateAjout }}</span>
+	<h1>{{ post.title }}</h1>
+	<span>Posté le {{ post.date }}</span>
 	<img src="{{ post.image }}">
-	<p>{{ post.contenu }}</p>
+	<p>{{ post.content }}</p>
 
 {% endblock %}
