--- core/Form/Validator/NotNullValidator.php
@@ -1,10 +1,0 @@
-<?php
-namespace Core\Form\Validator;
-
-class NotNullValidator extends Validator{
-
-	public function isValid($var){
-		return $var != '';		
-	}
-	
-}
