+++ core/Service/Router/Router.php
@@ -1,0 +1,22 @@
+<?php
+namespace Core\Service\Router;
+
+class Router {
+    
+    private $routes = array();
+    
+    public function route($pattern, $callback) {
+        $this->routes[$pattern] = $callback;
+    }
+    
+    public function execute($uri) {
+        foreach ($this->routes as $pattern => $callback) {
+            if (preg_match($pattern, $uri, $params) === 1) {
+                array_shift($params);
+                return call_user_func_array($callback, array_values($params));
+            }
+        }
+        throw new \Exception('No matching route');
+    }
+    
+}
