+++ core/Service/Validator/NotNull.php
@@ -1,0 +1,10 @@
+<?php
+namespace Core\Service\Validator;
+
+class NotNull extends Validator{
+
+	public function isValid($var){
+		return $var != '';		
+	}
+	
+}
